requirements.txt
watchdog>=3.0.0
pyyaml>=6.0.0
